requirements.txt
fastapi
azure.storage